feat: Update progress tracker and refactor component structure
