+++ docs/improvement-plan/progress-tracker.md
@@ -1,19 +1,19 @@
 # Master Progress Tracker
 
 ## 📊 Overall Project Status
-**Last Updated:** June 12, 2025  
-**Overall Progress:** 17.4% (53/305 total tasks completed)  
-**Active Phase:** Phase 1 (Testing Infrastructure)  
-**Current Focus:** Phase 1.2 (E2E Testing)
-**Next Milestone:** Complete Phase 1 testing infrastructure
+**Last Updated:** December 6, 2025  
+**Overall Progress:** 17.7% (54/305 total tasks completed)  
+**Active Phase:** Phase 2 (Frontend Code Quality & Structure)  
+**Current Focus:** Phase 2.1 (Modern React 19 Hooks Creation)
+**Next Milestone:** Complete Phase 2.1 Next.js 15 Application Architecture
 
 ## 🎯 Phase Overview & Progress
 
 | Phase | Status | Progress | Priority | Timeline | Dependencies |
 |-------|--------|----------|----------|----------|--------------|
 | [Phase 0: Foundation](./phase-0-foundation.md) | ✅ **Complete** | 100% (10/10) | Critical | 1-2 weeks | None |
-| [Phase 1: Testing](./phase-1-testing.md) | 🟡 **In Progress** | 40% (12/30) | Critical | 2-3 weeks | Phase 0 |
-| [Phase 2: Frontend](./phase-2-frontend.md) | ⭕ Not Started | 0% (0/25) | High | 3-4 weeks | Phase 1 |
+| [Phase 1: Testing](./phase-1-testing.md) | ✅ **Complete** | 100% (30/30) | Critical | 2-3 weeks | Phase 0 |
+| [Phase 2: Frontend](./phase-2-frontend.md) | 🟡 **In Progress** | 4% (1/25) | High | 3-4 weeks | Phase 1 |
 | [Phase 3: Performance](./phase-3-performance.md) | ⭕ Not Started | 0% (0/20) | High | 2-3 weeks | Phase 2 |
 | [Phase 4: Infrastructure](./phase-4-infrastructure.md) | ⭕ Not Started | 0% (0/18) | Medium | 2-3 weeks | Phase 3 |
 | [Phase 5: UX Design](./phase-5-ux-design.md) | ⭕ Not Started | 0% (0/22) | Medium | 3-4 weeks | Phase 2 |
@@ -23,32 +23,32 @@
 | [Quick Wins](./quick-wins.md) | ⭕ Not Started | 0% (0/15) | High | 1-2 weeks | None |
 
 **Total Tasks:** 305 across all phases  
-**Completed:** 53 tasks (17.4%)  
-**Remaining:** 252 tasks (82.6%)
+**Completed:** 54 tasks (17.7%)  
+**Remaining:** 251 tasks (82.3%)
 
 ## 📈 Progress Charts
 
 ### Phase Completion Status
 ```
 Phase 0: ████████████████████████████████████████  100% ✅
-Phase 1: ████████████████░░░░░░░░░░░░░░░░░░░░░░░░  40% 🟡
-Phase 2: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 3: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 4: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 5: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 6: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 7: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Phase 8: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
-Quick:   ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   0% ⭕
+Phase 1: ████████████████████████████████████████  100% ✅  
+Phase 2: ██░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    4% 🟡
+Phase 3: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Phase 4: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Phase 5: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Phase 6: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Phase 7: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Phase 8: ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
+Quick:   ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░    0% ⭕
 ```
 
 ### Priority Distribution
-- **Critical Priority:** 1 phase (Phase 1 at 40%)
-- **High Priority:** 4 phases (3 frontend-focused, 1 quick wins)
+- **Critical Priority:** 0 phases (Phase 1 complete)
+- **High Priority:** 4 phases (Phase 2 in progress, 2 pending, 1 quick wins)
 - **Medium Priority:** 3 phases (infrastructure, analytics, UX)
 - **Low Priority:** 1 phase (content strategy)
 
-## 🎯 Current Focus: Phase 1 Testing Infrastructure
+## 🎯 Current Focus: Phase 2 Frontend Code Quality & Structure
 
 ### Phase 0: Foundation Setup (Complete - 100% ✅)
