+++ static-web-app/__tests__/pages/blog-post.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import Page, {
-  generateStaticParams,
-  generateMetadata,
+    generateMetadata,
+    generateStaticParams,
 } from '../../app/blog/[slug]/page';
-import { ThemeProvider } from '../../components/ThemeProvider';
+import { ThemeProvider } from '../../components/providers/ThemeProvider';
 
 // Mock next/navigation
 jest.mock('next/navigation', () => ({
