+++ static-web-app/__tests__/pages/home.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import Page from '../../app/page';
-import { ThemeProvider } from '../../components/ThemeProvider';
+import { ThemeProvider } from '../../components/providers/ThemeProvider';
 
 // Mock next/navigation
 jest.mock('next/navigation', () => ({
