+++ static-web-app/app/blog/[slug]/page.tsx
@@ -1,12 +1,12 @@
 // --- file: app/blog/[slug]/page.tsx ---
-import { Metadata } from 'next';
-import { notFound } from 'next/navigation';
-import Link from 'next/link';
-import { MDXRemote } from 'next-mdx-remote/rsc';
-import { getAllPosts, getPostBySlug } from '@/lib/blog';
-import Container from '@/components/Container';
-import { formatDate } from '@/lib/utils';
 import { ArrowLeftIcon } from '@/components/icons/ArrowLeftIcon';
+import Container from '@/components/layout/Container';
+import { getAllPosts, getPostBySlug } from '@/lib/blog';
+import { formatDate } from '@/lib/utils';
+import { Metadata } from 'next';
+import { MDXRemote } from 'next-mdx-remote/rsc';
+import Link from 'next/link';
+import { notFound } from 'next/navigation';
 
 // Type for page props
 interface BlogPostPageProps {
