+++ static-web-app/app/contact/page.tsx
@@ -1,7 +1,7 @@
 // --- file: app/contact/page.tsx ---
-import { Metadata } from 'next';
-import Container from '@/components/Container';
 import { MailIcon } from '@/components/icons/MailIcon';
+import Container from '@/components/layout/Container';
+import { Metadata } from 'next';
 
 // SEO metadata for the contact page
 export const metadata: Metadata = {
