+++ static-web-app/app/layout.tsx
@@ -1,9 +1,9 @@
 // --- file: app/layout.tsx ---
+import { Layout } from '@/components/layout/Layout';
+import { ThemeProvider } from '@/components/providers/ThemeProvider';
 import type { Metadata } from 'next';
 import { Inter, JetBrains_Mono } from 'next/font/google';
 import './globals.css';
-import { Layout } from '@/components/Layout';
-import { ThemeProvider } from '@/components/ThemeProvider';
 
 // Configure Google Fonts
 const inter = Inter({
