+++ static-web-app/components/layout/Container.tsx
@@ -1,6 +1,6 @@
 // --- file: components/Container.tsx ---
-import { ReactNode, ElementType } from 'react';
-import { cn } from '../lib/utils';
+import { ElementType, ReactNode } from 'react';
+import { cn } from '../../lib/utils';
 
 interface ContainerProps {
   children: ReactNode;
