+++ static-web-app/components/navigation/MobileMenu.tsx
